Fixed a wrong explanation in 9.2
It said that you could go from a graph on a cylinder to a graph on a plane by cutting the cylinder along its height, which is not true if there are edges going all around the cylinder. I fixed it by giving another solution: projecting one end of the cylinder to a point and the other to infinity.